refac: move userIdToContext to different file
